pkg/endpoint: move LogStatusOKLocked and logStatusLocked to status.go
Signed-off by: Ian Vernon <ian@cilium.io>
