+++ pkg/endpoint/endpoint.go
@@ -34,7 +34,6 @@ import (
 	clientset "github.com/cilium/cilium/pkg/k8s/client/clientset/versioned"
 	pkgLabels "github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/lock"
-	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/mac"
 	"github.com/cilium/cilium/pkg/maps/policymap"
 	"github.com/cilium/cilium/pkg/option"
@@ -555,33 +554,6 @@ func ParseEndpoint(strEp string) (*Endpoint, error) {
 	ep.state = StateRestoring
 
 	return &ep, nil
-}
-
-// LogStatusOKLocked will log an OK message of the given status type with the
-// given msg string.
-// must be called with endpoint.Mutex held
-func (e *Endpoint) LogStatusOKLocked(typ StatusType, msg string) {
-	e.Status.indexMU.Lock()
-	defer e.Status.indexMU.Unlock()
-	e.logStatusLocked(typ, OK, msg)
-}
-
-func (e *Endpoint) logStatusLocked(typ StatusType, code StatusCode, msg string) {
-	sts := &statusLogMsg{
-		Status: Status{
-			Code:  code,
-			Msg:   msg,
-			Type:  typ,
-			State: e.state,
-		},
-		Timestamp: time.Now().UTC(),
-	}
-	e.Status.addStatusLog(sts)
-	e.getLogger().WithFields(logrus.Fields{
-		"code":                  sts.Status.Code,
-		"type":                  sts.Status.Type,
-		logfields.EndpointState: sts.Status.State,
-	}).Debug(msg)
 }
 
 type UpdateValidationError struct {
